Move DatasetEditor to QB components
