+++ frontend/src/metabase/query_builder/components/DatasetEditor/ResizableNotebook/ResizableNotebook.styled.jsx
