+++ frontend/src/metabase/query_builder/components/DatasetEditor/ResizableNotebook/index.js
